start mock styles

+++ calendar_enzyme/src/components/event/Event.jsx
@@ -1,31 +1,24 @@
-import React, { useState } from "react";
-import { useGlobalContext } from "../../context";
+import React, { useState, useEffect, useRef } from "react";
 import { MdDelete } from "react-icons/md";
+import { validateOnDelete } from "../../validation/validateModalInputs";
+import PropTypes from "prop-types";
 import "./event.scss";
 
-const Event = ({ id, height, marginTop, title, time, handleDelete, fetchEvents }) => {
-  const { isEvent, isOpen } = useGlobalContext();
+const Event = ({ id, height, marginTop, title, time, startTime, handleDelete, fetchEvents }) => {
+  const [isOpenDelete, setIsOpenDelete] = useState(false);
 
-  const [isClicked, setIsClicked] = useState(false);
+  const clickedHour = useRef();
 
-  const onOpenDelete = (e) => {
-    if (isOpen) {
-      return;
-    }
-    // if (e.target.className === 'event') {
-    //   return setIsClicked(!isEvent);
-    // }
-    if (!isEvent && !isClicked) {
-      return setIsClicked(!isEvent);
-    }
-    console.log(height);
-    return setIsClicked(!isClicked);
-  };
+  useOnClickOutside(clickedHour, () => setIsOpenDelete(false));
 
   function onCloseDelete() {
+    if (!validateOnDelete(startTime)) {
+      setIsOpenDelete(false);
+      return alert("You can do it 15 mins to event");
+    }
     handleDelete(id);
     fetchEvents();
-    setIsClicked(false);
+    setIsOpenDelete(false);
   }
 
   const eventStyle = {
@@ -36,14 +29,15 @@ const Event = ({ id, height, marginTop, title, time, handleDelete, fetchEvents }
   return (
     <>
       <div
+        ref={clickedHour}
         style={eventStyle}
         className="event"
-        onClick={onOpenDelete}
+        onClick={() => setIsOpenDelete(!isOpenDelete)}
         //
       >
         <div className="event__title">{title}</div>
         <div className="event__time">{time}</div>
-        {isClicked && (
+        {isOpenDelete && (
           <span
             className="delete-event-btn"
             style={{ top: height - 10 }}
@@ -58,4 +52,32 @@ const Event = ({ id, height, marginTop, title, time, handleDelete, fetchEvents }
   );
 };
 
+Event.propTypes = {
+  id: PropTypes.string,
+  height: PropTypes.number,
+  marginTop: PropTypes.number,
+  title: PropTypes.string,
+  time: PropTypes.string,
+  startTime: PropTypes.number,
+  handleDelete: PropTypes.func,
+  fetchEvents: PropTypes.func,
+};
 export default Event;
+
+function useOnClickOutside(ref, handler) {
+  useEffect(() => {
+    const listener = (event) => {
+      if (!ref.current || ref.current.contains(event.target)) {
+        return;
+      }
+
+      handler(event);
+    };
+
+    document.addEventListener("mousedown", listener);
+
+    return () => {
+      document.removeEventListener("mousedown", listener);
+    };
+  }, [ref, handler]);
+}
